data input module working
